+++ packages/core/src/random-settings.ts
@@ -921,6 +921,7 @@ export async function applyRandomSettings(rnd: OptionRandomSettings, oldSettings
       base.shuffleButterfliesOot = true;
       base.shuffleButterfliesMm = true;
       base.shuffleBouldersOot = 'all';
+      base.shuffleBouldersMm = 'all';
       base.shuffleRedBouldersOot = true;
       base.shuffleRedBouldersMm = true;
       base.shuffleSilverBouldersOot = true;
@@ -955,6 +956,7 @@ export async function applyRandomSettings(rnd: OptionRandomSettings, oldSettings
       base.shuffleButterfliesOot = true;
       base.shuffleButterfliesMm = true;
       base.shuffleBouldersOot = 'overworld';
+      base.shuffleBouldersMm = 'overworld';
       base.shuffleRedBouldersOot = true;
       base.shuffleRedBouldersMm = true;
       base.shuffleSilverBouldersOot = true;
@@ -980,6 +982,7 @@ export async function applyRandomSettings(rnd: OptionRandomSettings, oldSettings
       base.shuffleSnowballsMm = 'dungeons';
       base.shuffleWonderItemsOot = 'dungeons';
       base.shuffleBouldersOot = 'dungeons';
+      base.shuffleBouldersMm = 'dungeons';
       break;
     default:
       base.shufflePotsOot = sampleWeighted(random, { none: 10, overworld: 10, dungeons: 10, all: 10 });
@@ -1010,6 +1013,7 @@ export async function applyRandomSettings(rnd: OptionRandomSettings, oldSettings
       base.shuffleButterfliesOot = booleanWeighted(random, 0.5);
       base.shuffleButterfliesMm = booleanWeighted(random, 0.5);
       base.shuffleBouldersOot = sampleWeighted(random, { none: 10, overworld: 10, dungeons: 10, all: 10 });
+      base.shuffleBouldersMm = sampleWeighted(random, { none: 10, overworld: 10, dungeons: 10, all: 10 });
       base.shuffleRedBouldersOot = booleanWeighted(random, 0.5);
       base.shuffleRedBouldersMm = booleanWeighted(random, 0.5);
       base.shuffleSilverBouldersOot = booleanWeighted(random, 0.5);