+++ packages/logic/tests/seeds/extrasanity.test.ts
@@ -50,6 +50,7 @@ test("Can make a seed - Extrasanity", async () => {
     shuffleWonderItemsMm: true,
     fairySpotShuffleOot: true,
     shuffleBouldersOot: 'all',
+    shuffleBouldersMm: 'all',
     shuffleRedBouldersOot: true,
     shuffleRedBouldersMm: true,
     shuffleSilverBouldersOot: true,